# bridges/cloud-server/src/chat_sync/routes.rs
use std::sync::Arc;

use axum::extract::{Path, Query, State};
use axum::http::{HeaderMap, StatusCode};
use axum::middleware;
use axum::response::{IntoResponse, Response};
use axum::routing::{get, patch, post, put};
use axum::{Extension, Json, Router};
use serde_json::json;
use uuid::Uuid;

use crate::auth::rate_limit::{CloudRateLimiter, RateLimitDecision, MESSAGE_SEND_LIMIT};
use crate::auth::routes::{cloud_session_middleware, CloudSession};
use crate::chat_sync::cursor::CursorCodec;
use crate::chat_sync::models::{
    AddConversationMembersRequest, AdvanceConversationCursorRequest, BootstrapResponse,
    ConversationPreferencesResponse, ConversationResponse, CreateConversationRequest,
    CursorResponse, HistoryQuery, MessageResponse, RealtimeTicketResponse, SendMessageRequest,
    SyncQuery, SyncResponse, UpdateConversationTitleRequest, UpdatePersonalTitleRequest,
    UpdateReactionRequest,
};
use crate::chat_sync::realtime;
use crate::chat_sync::store::{self, StoreError};
use crate::chat_sync::PROTOCOL_VERSION;
use crate::server::ServerState;

const MAX_MESSAGE_CONTENT_BYTES: usize = 256 * 1024;
const MAX_ATTACHMENTS_PER_MESSAGE: usize = 32;
const MAX_IMAGE_PIXEL_DIMENSION: u64 = 100_000;

mod attachment_actions;
mod group_envelope;
mod http;
mod message_mutations;
mod reaction;
mod thread_reads;

use group_envelope::normalize_legacy_group_envelope;
use http::*;
#[derive(Clone)]
struct ChatSyncRuntime {
    cursor_codec: Option<CursorCodec>,
}

impl ChatSyncRuntime {
    fn from_env() -> Self {
        let cursor_codec = std::env::var("KORDI_CHAT_SYNC_CURSOR_SECRET")
            .ok()
            .and_then(|secret| CursorCodec::new(secret).ok());
        Self { cursor_codec }
    }
}

pub fn routes(state: Arc<ServerState>) -> Router {
    routes_with_runtime(state, ChatSyncRuntime::from_env())
}

fn routes_with_runtime(state: Arc<ServerState>, runtime: ChatSyncRuntime) -> Router {
    Router::new()
        .route("/v2/chat/conversations", post(create_conversation))
        .route(
            "/v2/chat/conversations/:conversation_id",
            patch(update_shared_title),
        )
        .route(
            "/v2/chat/conversations/:conversation_id/preferences",
            put(update_personal_title),
        )
        .route(
            "/v2/chat/conversations/:conversation_id/members",
            put(add_conversation_members),
        )
        .route(
            "/v2/chat/conversations/:conversation_id/messages",
            get(history).post(send_message),
        )
        .merge(message_mutations::routes())
        .merge(attachment_actions::routes())
        .route("/v2/chat/attention", get(thread_reads::attention))
        .route(
            "/v2/chat/conversations/:conversation_id/threads/:message_id",
            get(thread_reads::page),
        )
        .route(
            "/v2/chat/conversations/:conversation_id/threads/read",
            get(thread_reads::read).put(thread_reads::advance),
        )
        .route(
            "/v2/chat/conversations/:conversation_id/delivered",
            put(advance_delivery_cursor),
        )
        .route(
            "/v2/chat/conversations/:conversation_id/read",
            put(advance_read_cursor),
        )
        .route("/v2/chat/sync", get(sync))
        .route("/v2/chat/sync/bootstrap", get(bootstrap))
        .route("/v2/chat/realtime/ticket", post(issue_realtime_ticket))
        .layer(Extension(Arc::new(runtime)))
        .route_layer(middleware::from_fn_with_state(
            state.clone(),
            cloud_session_middleware,
        ))
        .with_state(state)
}

async fn create_conversation(
    State(state): State<Arc<ServerState>>,
    Extension(session): Extension<CloudSession>,
    Json(request): Json<CreateConversationRequest>,
) -> Response {
    let trusted_support_owner = state.support().and_then(|support| {
        let config = support.config();
        let expected_session_id = format!(
            "session:direct-system-agent:{}:{}",
            session.account_id, config.agent_id,
        );
        (request.kind == crate::chat_sync::models::ConversationKind::Direct
            && request.client_session_id.trim() == expected_session_id
            && request
                .member_account_ids
                .iter()
                .any(|member| member.trim() == config.owner_account_id))
        .then_some(config.owner_account_id.as_str())
    });
    match store::create_conversation_with_trusted_peer(
        state.db_pool(),
        &session.account_id,
        request,
        trusted_support_owner,
    )
    .await
    {
        Ok(outcome) => {
            if outcome.inserted {
                if let Some(pip) = state.pip() {
                    if let Err(error) = crate::pip::membership::join_conversation(
                        state.db_pool(),
                        &pip.config().account_id,
                        outcome.value.id,
                    )
                    .await
                    {
                        eprintln!("[pip] Could not join a new conversation: {error}");
                    }
                }
            }
            (
                if outcome.inserted {
                    StatusCode::CREATED
                } else {
                    StatusCode::OK
                },
                Json(ConversationResponse {
                    conversation: outcome.value,
                }),
            )
                .into_response()
        }
        Err(error) => store_error("create conversation", error),
    }
}

async fn send_message(
    State(state): State<Arc<ServerState>>,
    Extension(session): Extension<CloudSession>,
    rate_limiter: Option<Extension<Arc<CloudRateLimiter>>>,
    Path(conversation_id): Path<Uuid>,
    Json(mut request): Json<SendMessageRequest>,
) -> Response {
    if let Err(error) = normalize_legacy_group_envelope(&mut request) {
        return error.into_response();
    }
    if let Err(error) = validate_message_request(&request) {
        return error.into_response();
    }
    if let Some(Extension(rate_limiter)) = rate_limiter {
        if let RateLimitDecision::Limited { retry_after } = rate_limiter
            .observe_account_limit(MESSAGE_SEND_LIMIT, &session.account_id)
            .await
        {
            return rate_limited_response(retry_after);
        }
    }
    match store::send_message(
        state.db_pool(),
        &session.account_id,
        conversation_id,
        request,
    )
    .await
    {
        Ok(outcome) => {
            let message = match store::message_for_viewer(
                state.db_pool(),
                &session.account_id,
                outcome.value.clone(),
            )
            .await
            {
                Ok(message) => message,
                Err(error) => return store_error("project message", error),
            };
            if let Some(notifications) = state.notifications() {
                notifications
                    .send_message_attention(state.db_pool(), &outcome.value)
                    .await;
            }
            (
                if outcome.inserted {
                    StatusCode::CREATED
                } else {
                    StatusCode::OK
                },
                Json(MessageResponse { message }),
            )
                .into_response()
        }
        Err(error) => store_error("send message", error),
    }
}

async fn update_shared_title(
    State(state): State<Arc<ServerState>>,
    Extension(session): Extension<CloudSession>,
    Path(conversation_id): Path<Uuid>,
    Json(request): Json<UpdateConversationTitleRequest>,
) -> Response {
    match store::update_shared_title(
        state.db_pool(),
        &session.account_id,
        conversation_id,
        request,
    )
    .await
    {
        Ok(conversation) => Json(ConversationResponse { conversation }).into_response(),
        Err(error) => store_error("update shared title", error),
    }
}

async fn update_personal_title(
    State(state): State<Arc<ServerState>>,
    Extension(session): Extension<CloudSession>,
    Path(conversation_id): Path<Uuid>,
    Json(request): Json<UpdatePersonalTitleRequest>,
) -> Response {
    match store::update_personal_title(
        state.db_pool(),
        &session.account_id,
        conversation_id,
        request,
    )
    .await
    {
        Ok(preferences) => Json(ConversationPreferencesResponse { preferences }).into_response(),
        Err(error) => store_error("update personal title", error),
    }
}

async fn add_conversation_members(
    State(state): State<Arc<ServerState>>,
    Extension(session): Extension<CloudSession>,
    Path(conversation_id): Path<Uuid>,
    Json(request): Json<AddConversationMembersRequest>,
) -> Response {
    match store::add_conversation_members(
        state.db_pool(),
        &session.account_id,
        conversation_id,
        request,
    )
    .await
    {
        Ok(conversation) => Json(ConversationResponse { conversation }).into_response(),
        Err(error) => store_error("add conversation members", error),
    }
}

async fn history(
    State(state): State<Arc<ServerState>>,
    Extension(session): Extension<CloudSession>,
    Path(conversation_id): Path<Uuid>,
    Query(request): Query<HistoryQuery>,
) -> Response {
    match store::history(
        state.db_pool(),
        &session.account_id,
        conversation_id,
        request.before_sequence,
        request.limit,
    )
    .await
    {
        Ok(history) => Json(history).into_response(),
        Err(error) => store_error("load history", error),
    }
}

async fn advance_delivery_cursor(
    State(state): State<Arc<ServerState>>,
    Extension(session): Extension<CloudSession>,
    Path(conversation_id): Path<Uuid>,
    Json(request): Json<AdvanceConversationCursorRequest>,
) -> Response {
    match store::advance_delivery_cursor(
        state.db_pool(),
        &session.account_id,
        conversation_id,
        request,
    )
    .await
    {
        Ok(cursor) => Json(CursorResponse { cursor }).into_response(),
        Err(error) => store_error("advance delivery cursor", error),
    }
}

async fn advance_read_cursor(
    State(state): State<Arc<ServerState>>,
    Extension(session): Extension<CloudSession>,
    Path(conversation_id): Path<Uuid>,
    Json(request): Json<AdvanceConversationCursorRequest>,
) -> Response {
    match store::advance_read_cursor(
        state.db_pool(),
        &session.account_id,
        conversation_id,
        request,
    )
    .await
    {
        Ok(cursor) => Json(CursorResponse { cursor }).into_response(),
        Err(error) => store_error("advance read cursor", error),
    }
}

async fn sync(
    State(state): State<Arc<ServerState>>,
    Extension(runtime): Extension<Arc<ChatSyncRuntime>>,
    Extension(session): Extension<CloudSession>,
    Query(request): Query<SyncQuery>,
) -> Response {
    let codec = match require_cursor_codec(&runtime) {
        Ok(codec) => codec,
        Err(error) => return runtime_requirement_error(error),
    };
    let after_stream_seq = match request.cursor.as_deref() {
        Some(cursor) => match codec.decode(cursor, &session.account_id) {
            Ok(sequence) => sequence,
            Err(_) => {
                return error_response(
                    StatusCode::CONFLICT,
                    "SYNC_CURSOR_EXPIRED",
                    "The sync cursor is invalid for this account.",
                    Some(json!({ "bootstrap_required": true })),
                );
            }
        },
        None => 0,
    };
    match store::sync_batch(
        state.db_pool(),
        &session.account_id,
        after_stream_seq,
        request.limit,
    )
    .await
    {
        Ok(batch) => Json(SyncResponse {
            protocol_version: PROTOCOL_VERSION,
            content_removal_version: crate::chat_sync::content_removal_version(),
            events: batch.events,
            next_cursor: codec.encode(&session.account_id, batch.next_stream_seq),
            last_stream_seq: batch.next_stream_seq,
            has_more: batch.has_more,
            server_time: chrono::Utc::now(),
        })
        .into_response(),
        Err(error) => store_error("sync", error),
    }
}

async fn bootstrap(
    State(state): State<Arc<ServerState>>,
    Extension(runtime): Extension<Arc<ChatSyncRuntime>>,
    Extension(session): Extension<CloudSession>,
) -> Response {
    let codec = match require_cursor_codec(&runtime) {
        Ok(codec) => codec,
        Err(error) => return runtime_requirement_error(error),
    };
    match store::bootstrap(state.db_pool(), &session.account_id).await {
        Ok(snapshot) => Json(BootstrapResponse {
            session_visibility: snapshot.session_visibility,
            protocol_version: PROTOCOL_VERSION,
            content_removal_version: crate::chat_sync::content_removal_version(),
            conversations: snapshot.conversations,
            latest_messages: snapshot.latest_messages,
            session_pins: snapshot.session_pins,
            next_cursor: codec.encode(&session.account_id, snapshot.stream_seq),
            last_stream_seq: snapshot.stream_seq,
            server_time: snapshot.server_time,
        })
        .into_response(),
        Err(error) => store_error("bootstrap", error),
    }
}

async fn issue_realtime_ticket(
    State(state): State<Arc<ServerState>>,
    Extension(runtime): Extension<Arc<ChatSyncRuntime>>,
    Extension(session): Extension<CloudSession>,
    headers: HeaderMap,
) -> Response {
    if let Err(error) = require_cursor_codec(&runtime) {
        return runtime_requirement_error(error);
    }
    let origin = headers
        .get(axum::http::header::ORIGIN)
        .and_then(|value| value.to_str().ok());
    match realtime::issue_ticket(
        state.db_pool(),
        &session.account_id,
        &session.device_id,
        &session.token_id,
        origin,
    )
    .await
    {
        Ok(ticket) => (
            StatusCode::CREATED,
            Json(RealtimeTicketResponse {
                ticket: ticket.plaintext,
                device_id: session.device_id,
                expires_at: ticket.expires_at,
            }),
        )
            .into_response(),
        Err(realtime::TicketError::OriginNotAllowed) => error_response(
            StatusCode::FORBIDDEN,
            "REALTIME_ORIGIN_NOT_ALLOWED",
            "The browser origin is not allowed to open realtime connections.",
            None,
        ),
        Err(realtime::TicketError::Database(error)) => {
            eprintln!("[chat-sync] issue realtime ticket: {error}");
            error_response(
                StatusCode::INTERNAL_SERVER_ERROR,
                "SERVER_ERROR",
                "The realtime ticket could not be issued.",
                None,
            )
        }
        Err(realtime::TicketError::InvalidTicket) => error_response(
            StatusCode::BAD_REQUEST,
            "INVALID_REALTIME_TICKET",
            "The realtime ticket request is invalid.",
            None,
        ),
    }
}

#[cfg(test)]
mod tests;